requirements.txt
### install under venv python 3
--index-url https://pypi.python.org/simple
robotframework==6.0.2
requests==2.20.0
pyyaml==6.0
robotframework-appiumlibrary==2.0.0
robotframework-pabot==2.15.0
browserstack-sdk
